Shift event listeners to their own source files
This splits the event listeners for file events and key contexts out
into their own source files and away from the main override_audit.py
file.

These may eventually get merged back into a single external file, but
for the time being I've split them out due to some plans for making
more context handlers for key bindings, which (in my head, anyway)
seems like something that would be nicer to have in a separate file
for cleanliness.
