+++ src/__init__.py
@@ -1,10 +1,12 @@
 from ..bootstrap import reload
 
-reload("src", ["override_audit", "settings_proxy"])
+reload("src", ["override_audit", "events", "contexts", "settings_proxy"])
 reload("src.commands")
 
 from . import override_audit
 from .override_audit import *
+from .events import *
+from .contexts import *
 from .settings_proxy import *
 from .commands import *
 
@@ -25,7 +27,10 @@ __all__ = [
     "OverrideAuditContextOverrideCommand",
     "OverrideAuditContextPackageCommand",
     "OverrideAuditContextReportCommand",
+
+    # events/contexts
     "OverrideAuditEventListener",
+    "OverrideAuditContextListener",
 
     # commands/*
     "OverrideAuditModifyMarkCommand"
